\name{G_ind}
\alias{G_ind}
\docType{data}
\title{
K matrix of the group-wise independent kernel structure for the 9 ancestry groups from HMP3 data.}
\description{
K matrix of the group-wise independent kernel structure for the 9 ancestry groups from HMP3 data. The 9 ancestry groups are (in the following order): CEU - Utah residents with Northern and Western European ancestry from the CEPH collection; CHD - Chinese in Metropolitan Denver, Colorado; GIH - Gujarati Indians in Houston, Texas; JPT_CHB - Japanese in Tokyo, Japan & Han Chinese in Beijing, China; LWK - Luhya in Webuye, Kenya; MEX - Mexican ancestry in Los Angeles, California; MKK - Maasai in Kinyawa, Kenya; TSI - Toscani in Italia; YRI - Yoruba in Ibadan, Nigeria. }
\usage{data(G_ind)}
\details{
A 9 \eqn{\times} 9 diagonal matrix. The entries correspond to the order of the 9 ancestry groups as listed above.
}
\examples{

data(G_ind)
G_ind
}
\keyword{datasets}

