\name{TransMeta-package}
\alias{TransMeta-package}
\docType{package}
\title{
A novel random effect model for GWAS trans-ethnic meta-analysis 
}
\description{
This package implements the kernel machine based random effect model for testing single variant association in trans-ethnic meta-analysis
}
\details{
\tabular{ll}{
Package: \tab TransMeta\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2015-10-06\cr
License: \tab GPL (>= 2)\cr
}
This package implements the kernel machine based random effect model for single variant test in GWAS trans-ethnic Meta-analysis. In order to get p-value of the association test, the user should provide the effect-size estimates, standard error estimates and kernel matrix K in the Get_TransMeta function.}
\author{
Jingchunzi Shi, Seunggeun Lee

Maintainer: Jingchunzi (Jing) Shi <shijingc@umich.edu>

}
\keyword{ Effect size heterogeneity; Kernel regression; Random effect model; Trans-ethnic meta-analysis. }

